# src/frontend-angular/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": { "title": "AI Operations Orchestrator API", "version": "0.1.0" },
  "paths": {
...
    "/auth/users/{user_id}": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get user by ID",
        "description": "Get a user by ID (admin only).\n\nArgs:\n    user_id: User ID\n    current_user: Current authenticated admin user\n    db: Database session\n    \nReturns:\n    User information\n    \nRaises:\n    HTTPException: If user not found",
        "operationId": "get_user_auth_users__user_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Authentication"],
        "summary": "Update user",
        "description": "Update a user (admin only).\n\nArgs:\n    user_id: User ID\n    user_data: User update data\n    current_user: Current authenticated admin user\n    db: Database session\n    \nReturns:\n    Updated user information\n    \nRaises:\n    HTTPException: If user not found",
        "operationId": "update_user_auth_users__user_id__put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserResponse": {
        "properties": {
          "username": { "type": "string", "title": "Username" },
          "full_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "role": { "type": "string", "title": "Role", "default": "user" },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "user_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "User Metadata"
          },
          "id": { "type": "string", "title": "Id" },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Updated At"
          },
          "last_login": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Last Login"
          }
        },
        "type": "object",
        "required": ["username", "id"],
        "title": "UserResponse",
        "description": "Schema for user API responses."
      },
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "role": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Role"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          },
          "metadata": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Schema for updating a user."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}